requirements.txt
streamlit>=1.45.0
pandas
plotly
streamlit-aggrid
gspread
google-auth
